detect and report error

+++ pear/PEAR/Common.php
@@ -18,7 +18,7 @@
 // |                                                                      |
 // +----------------------------------------------------------------------+
 //
-// $Id: Common.php,v 1.14 2001/10/07 20:14:11 cox Exp $
+// $Id: Common.php,v 1.15 2001/10/08 03:31:14 cox Exp $
 
 require_once 'PEAR.php';
 require_once 'Archive/Tar.php';
@@ -386,7 +386,9 @@ class PEAR_Common extends PEAR
         $xml = $pkgdir . DIRECTORY_SEPARATOR . 'package.xml';
 
         $tar = new Archive_Tar($file, true);
-        $tar->extractList($xml);
+        if (!$tar->extractList($xml)) {
+            return $this->raiseError('could not extract the package.xml file');
+        }
         $info = $this->infoFromDescriptionFile($xml);
         unlink($xml);
         return $info;
